@@ -2,38 +2,32 @@
   "name": "@kit/eslint-config",
   "version": "0.2.0",
   "private": true,
-  "license": "MIT",
+  "type": "module",
   "files": [
-    "./apps.js"
+    "./apps.js",
+    "./base.js",
+    "./nextjs.js"
   ],
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
     "lint": "eslint .",
-    "format": "prettier --check \"**/*.{js,json}\"",
-    "typecheck": "tsc --noEmit"
+    "format": "prettier --check \"**/*.{js,json}\""
   },
   "dependencies": {
     "@next/eslint-plugin-next": "15.1.6",
-    "@types/eslint": "^8.56.10",
-    "@typescript-eslint/eslint-plugin": "^8.23.0",
-    "@typescript-eslint/parser": "^8.23.0",
+    "@types/eslint": "9.6.1",
+    "eslint-config-next": "15.1.6",
     "eslint-config-prettier": "^10.0.1",
     "eslint-config-turbo": "^2.4.0",
-    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-react": "7.37.4",
-    "eslint-plugin-react-hooks": "^5.1.0"
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "typescript-eslint": "8.23.0"
   },
   "devDependencies": {
     "@kit/prettier-config": "workspace:*",
     "@kit/tsconfig": "workspace:*",
-    "eslint": "^8.57.0",
+    "eslint": "^9.19.0",
     "typescript": "^5.7.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "./apps.js"
-    ]
   },
   "prettier": "@kit/prettier-config"
 }
