@@ -14,7 +14,7 @@ export function PostHeader({ post }: { post: Cms.ContentItem }) {
         <div className={'mx-auto flex max-w-3xl flex-col space-y-4'}>
           <h1
             className={
-              'font-heading text-3xl font-semibold tracking-tighter dark:text-white xl:text-5xl'
+              'font-heading text-3xl font-semibold tracking-tighter xl:text-5xl dark:text-white'
             }
           >
             {title}
@@ -27,7 +27,7 @@ export function PostHeader({ post }: { post: Cms.ContentItem }) {
           </div>
 
           <h2
-            className={'text-base text-muted-foreground xl:text-lg'}
+            className={'text-muted-foreground text-base xl:text-lg'}
             dangerouslySetInnerHTML={{ __html: description ?? '' }}
           ></h2>
         </div>
