@@ -9,7 +9,7 @@ export const CtaButton: React.FC<React.ComponentProps<typeof Button>> =
           'h-12 rounded-xl px-4 text-base font-semibold',
           className,
           {
-            ['transition-all hover:shadow-2xl dark:shadow-primary/30']:
+            ['dark:shadow-primary/30 transition-all hover:shadow-2xl']:
               props.variant === 'default' || !props.variant,
           },
         )}
