@@ -34,7 +34,7 @@ const AvatarFallback: React.FC<
 > = ({ className, ...props }) => (
   <AvatarPrimitive.Fallback
     className={cn(
-      'flex h-full w-full items-center justify-center rounded-full bg-muted',
+      'bg-muted flex h-full w-full items-center justify-center rounded-full',
       className,
     )}
     {...props}
