@@ -18,7 +18,7 @@ const BreadcrumbList: React.FC<React.ComponentPropsWithRef<'ol'>> = ({
 }) => (
   <ol
     className={cn(
-      'flex flex-wrap items-center gap-1.5 break-words text-sm text-muted-foreground',
+      'text-muted-foreground flex flex-wrap items-center gap-1.5 text-sm break-words',
       className,
     )}
     {...props}
@@ -64,7 +64,7 @@ const BreadcrumbPage: React.FC<React.ComponentPropsWithoutRef<'span'>> = ({
     role="link"
     aria-disabled="true"
     aria-current="page"
-    className={cn('font-normal text-foreground', className)}
+    className={cn('text-foreground font-normal', className)}
     {...props}
   />
 );
