@@ -12,7 +12,7 @@ export function Heading({
       return (
         <h1
           className={cn(
-            `scroll-m-20 font-heading text-3xl font-bold tracking-tight dark:text-white lg:text-4xl`,
+            `font-heading scroll-m-20 text-3xl font-bold tracking-tight lg:text-4xl dark:text-white`,
             className,
           )}
         >
@@ -23,7 +23,7 @@ export function Heading({
       return (
         <h2
           className={cn(
-            `scroll-m-20 pb-2 font-heading text-2xl font-semibold tracking-tight transition-colors first:mt-0 lg:text-3xl`,
+            `font-heading scroll-m-20 pb-2 text-2xl font-semibold tracking-tight transition-colors first:mt-0 lg:text-3xl`,
             className,
           )}
         >
@@ -34,7 +34,7 @@ export function Heading({
       return (
         <h3
           className={cn(
-            'scroll-m-20 font-heading text-xl font-semibold tracking-tight lg:text-2xl',
+            'font-heading scroll-m-20 text-xl font-semibold tracking-tight lg:text-2xl',
             className,
           )}
         >
@@ -45,7 +45,7 @@ export function Heading({
       return (
         <h4
           className={cn(
-            'scroll-m-20 font-heading text-lg font-semibold tracking-tight lg:text-xl',
+            'font-heading scroll-m-20 text-lg font-semibold tracking-tight lg:text-xl',
             className,
           )}
         >
@@ -56,7 +56,7 @@ export function Heading({
       return (
         <h5
           className={cn(
-            'scroll-m-20 font-heading text-base font-medium lg:text-lg',
+            'font-heading scroll-m-20 text-base font-medium lg:text-lg',
             className,
           )}
         >
@@ -67,7 +67,7 @@ export function Heading({
       return (
         <h6
           className={cn(
-            'scroll-m-20 font-heading text-base font-medium',
+            'font-heading scroll-m-20 text-base font-medium',
             className,
           )}
         >
