@@ -11,13 +11,13 @@ const Progress: React.FC<
 > = ({ className, value, ...props }) => (
   <ProgressPrimitive.Root
     className={cn(
-      'relative h-2 w-full overflow-hidden rounded-full bg-primary/20',
+      'bg-primary/20 relative h-2 w-full overflow-hidden rounded-full',
       className,
     )}
     {...props}
   >
     <ProgressPrimitive.Indicator
-      className="h-full w-full flex-1 bg-primary transition-all"
+      className="bg-primary h-full w-full flex-1 transition-all"
       style={{ transform: `translateX(-${100 - (value ?? 0)}%)` }}
     />
   </ProgressPrimitive.Root>
